# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "URL Shortener API",
    "version": "1.0.0",
    "description": "Documentation of the URL Shortener API.",
    "contact": {
      "name": "Jacson Rodrigues",
      "emai": "contact.jacsonrodrigues@proton.me",
      "url": "https://www.linkedin.com/in/jacson-rodrigues/"
    }
  },
  "host": "localhost:8080",
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authentication",
        "description": "Authentication endpoint for user login.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Login Data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "POST /users",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "GET /users",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/users/me": {
      "get": {
        "summary": "GET /users/me",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "PATCH /users/me",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "DELETE /users/me",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/shorten/auth": {
      "post": {
        "summary": "POST /shorten/auth",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/shorten": {
      "post": {
        "summary": "POST /shorten",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": []
      }
    },
    "/me/urls": {
      "get": {
        "summary": "GET /me/urls",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/me/urls/{id}": {
      "patch": {
        "summary": "PATCH /me/urls/{id}",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "DELETE /me/urls/{id}",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{shortCode}": {
      "get": {
        "summary": "GET /{shortCode}",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Url": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "shortCode": {
          "type": "string",
          "maxLength": 6,
          "pattern": "^[-a-zA-Z0-9]{1,6}$"
        },
        "originalUrl": {
          "type": "string"
        },
        "userId": {
          "type": "string",
          "format": "uuid"
        },
        "clicks": {
          "type": "integer",
          "minimum": 0
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "deletedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "deletedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "UserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CreateUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "format": "password"
        },
        "name": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "UpdateUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "format": "password"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "AuthResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "format": "jwt"
        }
      }
    },
    "LoginDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "format": "password"
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "KrakenD Gateway"
    }
  ]
}